rerunsdjango/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
backports-zoneinfo==0.2.1
    # via
    #   django
    #   django-celery-beat
    #   django-timezone-field
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2022.12.7
    # via
    #   requests
    #   rssreruns
charset-normalizer==3.0.1
    # via
    #   requests
    #   rssreruns
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
django==4.1
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-invitations
django-celery-beat==2.4.0
    # via -r requirements.in
django-invitations==2.0.0
    # via -r requirements.in
django-timezone-field==5.0
    # via
    #   -r requirements.in
    #   django-celery-beat
idna==3.4
    # via
    #   requests
    #   rssreruns
kombu==5.2.4
    # via celery
lxml==4.9.2
    # via rssreruns
prompt-toolkit==3.0.36
    # via click-repl
psycopg2-binary==2.9.5
    # via -r requirements.in
python-crontab==2.7.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   python-crontab
    #   rssreruns
pytz==2022.7.1
    # via
    #   celery
    #   django-timezone-field
redis==4.4.2
    # via -r requirements.in
requests==2.28.2
    # via rssreruns
rssreruns==0.0.4
    # via -r requirements.in
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
    #   rssreruns
sqlparse==0.4.3
    # via django
supervisor==4.2.5
    # via -r requirements.in
tzdata==2022.7
    # via django-celery-beat
urllib3==1.26.14
    # via
    #   requests
    #   rssreruns
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.1.9
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
